pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asciibench"
version = "0.1.0"
description = "Elo-based LLM benchmark system with Generator, Judge UI, and Analyst modules"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "smolagents[litellm]",
    "python-multipart",
    "rich>=14.3.1",
    "logfire>=4.21.0",
    "filelock>=3.20.3",
    "pillow>=10.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio>=0.23.0",
    "pytest-cov",
    "ruff",
    "ty",
    "pre-commit",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PT", "RUF"]
ignore = ["UP047", "UP046"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

# Tailwind CSS is installed via npm (package.json)
# Build with: npm run build:css
# Watch with: npm run watch:css
